refactor

+++ cpp/WMD.h
@@ -19,8 +19,8 @@ public:
   void save(FILE * fout) const;
   void load(FILE * fin);
   real rwmd(WeightedDocument * src, UnWeightedDocument * target);
-  void sent_knn_docs(TaggedDocument & doc, knn_item_t * knns, int k);
-  void sent_knn_docs_ex(TaggedDocument & doc, knn_item_t * knns, int k);
+  void sent_knn_docs(TaggedDocument & doc, knn_item_t * knns, size_t k);
+  void sent_knn_docs_ex(TaggedDocument & doc, knn_item_t * knns, size_t k);
 
 private:
   void loadFromDoc2Vec();
